+++ secure-iframe/src/ui.ts
@@ -33,7 +33,7 @@ export function createInputElement(type: IframeFieldType) {
 
   switch (type) {
     case "credit_card_number": {
-      input.placeholder = "4000 0000 0000 0000";
+      input.placeholder = "1234 1234 1234 1234";
       input.inputMode = "numeric";
       input.autocomplete = "cc-number";
       input.maxLength = 23; // 19 digits + 4 spaces
@@ -309,6 +309,7 @@ export function createFatalErrorComponent(code: string) {
   div.style.color = "red";
   div.style.fontFamily = "monospace";
   div.style.fontSize = "10px";
+  div.style.margin = "12px";
   div.textContent = `✕ ${code}`;
   return div;
 }